Remove StorageClusterInitialization
This commit removes the need for a StorageClusterInitialization CR by
replacing it with a set of strings to controll the reconciliation of
previously initialized CRs. This changes the behavior for reapplying the
initialization, requiring the admin to set each resource's strategy to
"manage", wait for reconciliation, then back to "" or "default".

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>